have to set all=T, not all.x=T to ensure we have all countries
e.g. even those that happen not to impose any of the interventions under "inward subsidies" and hence would be absent from the first table

This should do the trick here since we are talking about well-monitored G20 nations, but the cleanest way to do this is to construct a data.frame with all the countries in it and then merge onto that.